requirements.txt
fastapi
motor
python-decouple
pymongo
uvicorn